+++ src/Sk/SmartId/Exception/InvalidParametersException.php
@@ -26,6 +26,6 @@
  */
 namespace Sk\SmartId\Exception;
 
-class InvalidParametersException extends SmartIdException
+class InvalidParametersException extends EnduringSmartIdException
 {
 }