+++ src/Sk/SmartId/Exception/ServerMaintenanceException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Sk\SmartId\Exception;
+
+// Server is under maintenance, retry later.
+class ServerMaintenanceException extends EnduringSmartIdException
+{
+
+}