+++ src/Sk/SmartId/Exception/UserRefusedConfirmationMessageWithVcChoiceException.php
@@ -4,7 +4,7 @@
 namespace Sk\SmartId\Exception;
 
 
-class UserRefusedConfirmationMessageWithVcChoiceException extends SmartIdException
+class UserRefusedConfirmationMessageWithVcChoiceException extends UserRefusedException
 {
 
 }