+++ src/Sk/SmartId/Exception/UserSelectedWrongVerificationCodeException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Sk\SmartId\Exception;
+
+class UserSelectedWrongVerificationCodeException extends UserActionException
+{
+
+}